Move visit operations to VisitService (#577)
