+++ integration_tests/mockApis/orchestration.ts
@@ -1,9 +1,94 @@
 import { SuperAgentRequest } from 'superagent'
 import { stubFor } from './wiremock'
-import { PrisonerProfile, VisitHistoryDetails } from '../../server/data/orchestrationApiTypes'
+import { OutcomeDto, PrisonerProfile, Visit, VisitHistoryDetails } from '../../server/data/orchestrationApiTypes'
 import TestData from '../../server/routes/testutils/testData'
 
 export default {
+  stubBookVisit: (visit: Visit): SuperAgentRequest => {
+    return stubFor({
+      request: {
+        method: 'PUT',
+        url: `/orchestration/visits/${visit.applicationReference}/book`,
+      },
+      response: {
+        status: 200,
+        headers: { 'Content-Type': 'application/json;charset=UTF-8' },
+        jsonBody: visit,
+      },
+    })
+  },
+  stubCancelVisit: ({ visit, outcome }: { visit: Visit; outcome: OutcomeDto }): SuperAgentRequest => {
+    return stubFor({
+      request: {
+        method: 'PUT',
+        url: `/orchestration/visits/${visit.reference}/cancel`,
+        bodyPatterns: [
+          {
+            equalToJson: {
+              outcomeStatus: outcome.outcomeStatus,
+              text: outcome.text,
+            },
+            ignoreArrayOrder: true,
+          },
+        ],
+      },
+      response: {
+        status: 200,
+        headers: { 'Content-Type': 'application/json;charset=UTF-8' },
+        jsonBody: visit,
+      },
+    })
+  },
+  stubChangeReservedSlot: (visit: Visit): SuperAgentRequest => {
+    return stubFor({
+      request: {
+        method: 'PUT',
+        url: `/orchestration/visits/${visit.applicationReference}/slot/change`,
+        bodyPatterns: [
+          {
+            equalToJson: {
+              visitRestriction: visit.visitRestriction,
+              startTimestamp: visit.startTimestamp,
+              endTimestamp: visit.endTimestamp,
+              visitContact: visit.visitContact,
+              visitors: visit.visitors,
+              visitorSupport: visit.visitorSupport,
+            },
+          },
+        ],
+      },
+      response: {
+        status: 200,
+        headers: { 'Content-Type': 'application/json;charset=UTF-8' },
+        jsonBody: visit,
+      },
+    })
+  },
+  stubReserveVisit: (visit: Visit): SuperAgentRequest => {
+    return stubFor({
+      request: {
+        method: 'POST',
+        url: '/orchestration/visits/slot/reserve',
+        bodyPatterns: [
+          {
+            equalToJson: {
+              prisonerId: visit.prisonerId,
+              sessionTemplateReference: visit.sessionTemplateReference,
+              visitRestriction: visit.visitRestriction,
+              startTimestamp: visit.startTimestamp,
+              endTimestamp: visit.endTimestamp,
+              visitors: visit.visitors,
+            },
+          },
+        ],
+      },
+      response: {
+        status: 200,
+        headers: { 'Content-Type': 'application/json;charset=UTF-8' },
+        jsonBody: visit,
+      },
+    })
+  },
   stubVisitHistory: (visitHistoryDetails: VisitHistoryDetails): SuperAgentRequest => {
     return stubFor({
       request: {
