+++ server/routes/bookAVisit.ts
@@ -16,6 +16,7 @@ export default function routes({
   notificationsService,
   prisonerProfileService,
   prisonerVisitorsService,
+  visitService,
   visitSessionsService,
 }: Services): Router {
   const router = Router()
@@ -35,9 +36,9 @@ export default function routes({
   const selectVisitors = new SelectVisitors('book', prisonerVisitorsService, prisonerProfileService)
   const visitType = new VisitType('book', auditService)
   const additionalSupport = new AdditionalSupport('book', additionalSupportService)
-  const dateAndTime = new DateAndTime('book', visitSessionsService, auditService)
+  const dateAndTime = new DateAndTime('book', visitService, visitSessionsService, auditService)
   const mainContact = new MainContact('book')
-  const checkYourBooking = new CheckYourBooking('book', visitSessionsService, auditService, notificationsService)
+  const checkYourBooking = new CheckYourBooking('book', auditService, notificationsService, visitService)
   const confirmation = new Confirmation('book')
 
   get('/select-visitors', sessionCheckMiddleware({ stage: 1 }), (req, res) => selectVisitors.get(req, res))
