+++ server/routes/visitJourney/dateAndTime.ts
@@ -2,13 +2,14 @@ import type { Request, Response } from 'express'
 import { body, ValidationChain, validationResult } from 'express-validator'
 import { VisitSlot } from '../../@types/bapv'
 import AuditService from '../../services/auditService'
-import VisitSessionsService from '../../services/visitSessionsService'
 import { getFlashFormValues, getSelectedSlot, getSlotByTimeAndRestriction } from '../visitorUtils'
 import getUrlPrefix from './visitJourneyUtils'
+import { VisitService, VisitSessionsService } from '../../services'
 
 export default class DateAndTime {
   constructor(
     private readonly mode: string,
+    private readonly visitService: VisitService,
     private readonly visitSessionsService: VisitSessionsService,
     private readonly auditService: AuditService,
   ) {}
@@ -109,12 +110,12 @@ export default class DateAndTime {
 
     // See README ('Visit journeys – book and update') for explanation of this flow
     if (visitSessionData.applicationReference) {
-      await this.visitSessionsService.changeReservedVisit({
+      await this.visitService.changeReservedVisit({
         username: res.locals.user.username,
         visitSessionData,
       })
     } else if (isUpdate) {
-      const { applicationReference, visitStatus } = await this.visitSessionsService.changeBookedVisit({
+      const { applicationReference, visitStatus } = await this.visitService.changeBookedVisit({
         username: res.locals.user.username,
         visitSessionData,
       })
@@ -122,7 +123,7 @@ export default class DateAndTime {
       visitSessionData.applicationReference = applicationReference
       visitSessionData.visitStatus = visitStatus
     } else {
-      const { applicationReference, reference, visitStatus } = await this.visitSessionsService.reserveVisit({
+      const { applicationReference, reference, visitStatus } = await this.visitService.reserveVisit({
         username: res.locals.user.username,
         visitSessionData,
       })
