+++ server/services/visitService.ts
@@ -1,5 +1,5 @@
-import { VisitorListItem } from '../@types/bapv'
-import { VisitHistoryDetails } from '../data/orchestrationApiTypes'
+import { VisitSessionData, VisitorListItem } from '../@types/bapv'
+import { OutcomeDto, Visit, VisitHistoryDetails } from '../data/orchestrationApiTypes'
 import buildVisitorListItem from '../utils/visitorUtils'
 import { getSupportTypeDescriptions } from '../routes/visitorUtils'
 import { HmppsAuthClient, OrchestrationApiClient, PrisonerContactRegistryApiClient, RestClientBuilder } from '../data'
@@ -12,6 +12,77 @@ export default class VisitService {
     private readonly additionalSupportService: AdditionalSupportService,
     private readonly hmppsAuthClient: HmppsAuthClient,
   ) {}
+
+  async bookVisit({
+    username,
+    applicationReference,
+  }: {
+    username: string
+    applicationReference: string
+  }): Promise<Visit> {
+    const token = await this.hmppsAuthClient.getSystemClientToken(username)
+    const orchestrationApiClient = this.orchestrationApiClientFactory(token)
+
+    const visit = await orchestrationApiClient.bookVisit(applicationReference)
+    return visit
+  }
+
+  async cancelVisit({
+    username,
+    reference,
+    outcome,
+  }: {
+    username: string
+    reference: string
+    outcome: OutcomeDto
+  }): Promise<Visit> {
+    const token = await this.hmppsAuthClient.getSystemClientToken(username)
+    const orchestrationApiClient = this.orchestrationApiClientFactory(token)
+
+    return orchestrationApiClient.cancelVisit(reference, outcome)
+  }
+
+  async changeBookedVisit({
+    username,
+    visitSessionData,
+  }: {
+    username: string
+    visitSessionData: VisitSessionData
+  }): Promise<Visit> {
+    const token = await this.hmppsAuthClient.getSystemClientToken(username)
+    const orchestrationApiClient = this.orchestrationApiClientFactory(token)
+
+    const visit = await orchestrationApiClient.changeBookedVisit(visitSessionData)
+    return visit
+  }
+
+  async changeReservedVisit({
+    username,
+    visitSessionData,
+  }: {
+    username: string
+    visitSessionData: VisitSessionData
+  }): Promise<Visit> {
+    const token = await this.hmppsAuthClient.getSystemClientToken(username)
+    const orchestrationApiClient = this.orchestrationApiClientFactory(token)
+
+    const visit = await orchestrationApiClient.changeReservedVisit(visitSessionData)
+    return visit
+  }
+
+  async reserveVisit({
+    username,
+    visitSessionData,
+  }: {
+    username: string
+    visitSessionData: VisitSessionData
+  }): Promise<Visit> {
+    const token = await this.hmppsAuthClient.getSystemClientToken(username)
+    const orchestrationApiClient = this.orchestrationApiClientFactory(token)
+
+    const reservation = await orchestrationApiClient.reserveVisit(visitSessionData)
+    return reservation
+  }
 
   async getFullVisitDetails({
     username,
