+++ server/services/visitSessionsService.test.ts
@@ -1,13 +1,7 @@
 import { NotFound } from 'http-errors'
-import {
-  VisitSlotList,
-  VisitSessionData,
-  VisitInformation,
-  ExtendedVisitInformation,
-  VisitsPageSlot,
-} from '../@types/bapv'
+import { VisitSlotList, VisitInformation, ExtendedVisitInformation, VisitsPageSlot } from '../@types/bapv'
 import { Address, Contact, AddressUsage, Restriction } from '../data/prisonerContactRegistryApiTypes'
-import { VisitSession, Visit, OutcomeDto, PageVisitDto, SessionSchedule } from '../data/orchestrationApiTypes'
+import { VisitSession, Visit, PageVisitDto, SessionSchedule } from '../data/orchestrationApiTypes'
 import { ScheduledEvent } from '../data/whereaboutsApiTypes'
 import TestData from '../routes/testutils/testData'
 import VisitSessionsService from './visitSessionsService'
@@ -679,331 +673,6 @@ describe('Visit sessions service', () => {
         sessionEndTime,
       )
       expect(results).toEqual(sessionCapacity)
-    })
-  })
-
-  describe('reserveVisitSlot', () => {
-    it('should create a new RESERVED visit and return the visit data', async () => {
-      const visitSessionData: VisitSessionData = {
-        prisoner: {
-          offenderNo: 'A1234BC',
-          name: 'prisoner name',
-          dateOfBirth: '23 May 1988',
-          location: 'somewhere',
-        },
-        visitSlot: {
-          id: '1',
-          sessionTemplateReference: 'v9d.7ed.7u',
-          prisonId,
-          startTimestamp: '2022-02-14T10:00:00',
-          endTimestamp: '2022-02-14T11:00:00',
-          availableTables: 1,
-          capacity: 15,
-          visitRoom: 'visit room',
-          visitRestriction: 'OPEN',
-        },
-        visitRestriction: 'OPEN',
-        visitors: [
-          {
-            personId: 123,
-            name: 'visitor name',
-            relationshipDescription: 'relationship desc',
-            restrictions: [
-              {
-                restrictionType: 'TEST',
-                restrictionTypeDescription: 'test type',
-                startDate: '10 May 2020',
-                expiryDate: '10 May 2022',
-                globalRestriction: false,
-                comment: 'comments',
-              },
-            ],
-            banned: false,
-          },
-        ],
-      }
-      const visit: Visit = {
-        applicationReference: 'aaa-bbb-ccc',
-        reference: 'ab-cd-ef-gh',
-        prisonerId: visitSessionData.prisoner.offenderNo,
-        prisonId: 'HEI',
-        sessionTemplateReference: 'v9d.7ed.7u',
-        visitRoom: visitSessionData.visitSlot.visitRoom,
-        visitType: 'SOCIAL',
-        visitStatus: 'RESERVED',
-        visitRestriction: 'OPEN',
-        startTimestamp: '2022-02-14T10:00:00',
-        endTimestamp: '2022-02-14T11:00:00',
-        visitNotes: [],
-        visitors: [
-          {
-            nomisPersonId: 1234,
-          },
-        ],
-        visitorSupport: [],
-        createdBy: 'user1',
-        createdTimestamp: '2022-02-14T10:00:00',
-        modifiedTimestamp: '2022-02-14T10:05:00',
-      }
-
-      visitSchedulerApiClient.reserveVisit.mockResolvedValue(visit)
-      whereaboutsApiClient.getEvents.mockResolvedValue([])
-      const result = await visitSessionsService.reserveVisit({ username: 'user', visitSessionData })
-
-      expect(visitSchedulerApiClient.reserveVisit).toHaveBeenCalledTimes(1)
-      expect(result).toEqual(visit)
-    })
-  })
-
-  describe('changeReservedVisit', () => {
-    it('should change an existing reserved visit and return the visit data', async () => {
-      const visitSessionData: VisitSessionData = {
-        prisoner: {
-          offenderNo: 'A1234BC',
-          name: 'prisoner name',
-          dateOfBirth: '23 May 1988',
-          location: 'somewhere',
-        },
-        visitSlot: {
-          id: '1',
-          sessionTemplateReference: 'v9d.7ed.7u',
-          prisonId,
-          startTimestamp: '2022-02-14T10:00:00',
-          endTimestamp: '2022-02-14T11:00:00',
-          availableTables: 1,
-          capacity: 15,
-          visitRoom: 'visit room',
-          visitRestriction: 'OPEN',
-        },
-        visitRestriction: 'OPEN',
-        visitors: [
-          {
-            personId: 123,
-            name: 'visitor name',
-            relationshipDescription: 'rel desc',
-            restrictions: [
-              {
-                restrictionType: 'TEST',
-                restrictionTypeDescription: 'test type',
-                startDate: '10 May 2020',
-                expiryDate: '10 May 2022',
-                globalRestriction: false,
-                comment: 'comments',
-              },
-            ],
-            banned: false,
-          },
-        ],
-        visitorSupport: [{ type: 'WHEELCHAIR' }, { type: 'MASK_EXEMPT' }, { type: 'OTHER', text: 'custom request' }],
-        mainContact: {
-          phoneNumber: '01234 567890',
-          contactName: 'John Smith',
-        },
-        applicationReference: 'aaa-bbb-ccc',
-        visitReference: 'ab-cd-ef-gh',
-        visitStatus: 'RESERVED',
-      }
-
-      const visit: Visit = {
-        applicationReference: visitSessionData.applicationReference,
-        reference: visitSessionData.visitReference,
-        prisonerId: visitSessionData.prisoner.offenderNo,
-        prisonId: 'HEI',
-        sessionTemplateReference: 'v9d.7ed.7u',
-        visitRoom: visitSessionData.visitSlot.visitRoom,
-        visitType: 'SOCIAL',
-        visitStatus: visitSessionData.visitStatus,
-        visitRestriction: visitSessionData.visitRestriction,
-        startTimestamp: visitSessionData.visitSlot.startTimestamp,
-        endTimestamp: visitSessionData.visitSlot.endTimestamp,
-        visitNotes: [],
-        visitContact: {
-          name: 'John Smith',
-          telephone: '01234 567890',
-        },
-        visitors: [
-          {
-            nomisPersonId: 1234,
-          },
-        ],
-        visitorSupport: [{ type: 'WHEELCHAIR' }, { type: 'MASK_EXEMPT' }, { type: 'OTHER', text: 'custom request' }],
-        createdBy: 'user1',
-        createdTimestamp: '2022-02-14T10:00:00',
-        modifiedTimestamp: '2022-02-14T10:05:00',
-      }
-
-      visitSchedulerApiClient.changeReservedVisit.mockResolvedValue(visit)
-      whereaboutsApiClient.getEvents.mockResolvedValue([])
-
-      const result = await visitSessionsService.changeReservedVisit({
-        username: 'user',
-        visitSessionData,
-      })
-
-      expect(visitSchedulerApiClient.changeReservedVisit).toHaveBeenCalledTimes(1)
-      expect(result).toEqual(visit)
-    })
-  })
-
-  describe('bookVisit', () => {
-    it('should book a visit (change status from RESERVED to BOOKED)', async () => {
-      const applicationReference = 'aaa-bbb-ccc'
-
-      const visit: Partial<Visit> = {
-        applicationReference,
-        reference: 'ab-cd-ef-gh',
-        visitStatus: 'BOOKED',
-      }
-
-      visitSchedulerApiClient.bookVisit.mockResolvedValue(visit as Visit)
-      whereaboutsApiClient.getEvents.mockResolvedValue([])
-      const result = await visitSessionsService.bookVisit({ username: 'user', applicationReference })
-
-      expect(visitSchedulerApiClient.bookVisit).toHaveBeenCalledTimes(1)
-      expect(result).toEqual(visit)
-    })
-  })
-
-  describe('changeBookedVisit', () => {
-    it('should change an existing booked visit and return the visit data with status RESERVED/CHANGING and new applicationReference', async () => {
-      const visitSessionData: VisitSessionData = {
-        prisoner: {
-          offenderNo: 'A1234BC',
-          name: 'prisoner name',
-          dateOfBirth: '23 May 1988',
-          location: 'somewhere',
-        },
-        visitSlot: {
-          id: '1',
-          sessionTemplateReference: 'v9d.7ed.7u',
-          prisonId,
-          startTimestamp: '2022-02-14T10:00:00',
-          endTimestamp: '2022-02-14T11:00:00',
-          availableTables: 1,
-          capacity: 15,
-          visitRoom: 'visit room',
-          visitRestriction: 'OPEN',
-        },
-        visitRestriction: 'OPEN',
-        visitors: [
-          {
-            personId: 123,
-            name: 'visitor name',
-            relationshipDescription: 'relationship desc',
-            restrictions: [
-              {
-                restrictionType: 'TEST',
-                restrictionTypeDescription: 'test type',
-                startDate: '10 May 2020',
-                expiryDate: '10 May 2022',
-                globalRestriction: false,
-                comment: 'comments',
-              },
-            ],
-            banned: false,
-          },
-        ],
-        visitorSupport: [{ type: 'WHEELCHAIR' }],
-        mainContact: {
-          phoneNumber: '01234 567890',
-          contactName: 'John Smith',
-        },
-        applicationReference: 'aaa-bbb-ccc',
-        visitReference: 'ab-cd-ef-gh',
-        visitStatus: 'BOOKED',
-      }
-
-      const returnedVisit: Visit = {
-        applicationReference: 'ddd-eee-fff',
-        reference: visitSessionData.visitReference,
-        prisonerId: visitSessionData.prisoner.offenderNo,
-        prisonId: 'HEI',
-        sessionTemplateReference: 'v9d.7ed.7u',
-        visitRoom: visitSessionData.visitSlot.visitRoom,
-        visitType: 'SOCIAL',
-        visitStatus: 'CHANGING',
-        visitRestriction: visitSessionData.visitRestriction,
-        startTimestamp: visitSessionData.visitSlot.startTimestamp,
-        endTimestamp: visitSessionData.visitSlot.endTimestamp,
-        visitNotes: [],
-        visitContact: {
-          name: 'John Smith',
-          telephone: '01234 567890',
-        },
-        visitors: [
-          {
-            nomisPersonId: 1234,
-          },
-        ],
-        visitorSupport: [{ type: 'WHEELCHAIR' }],
-        createdBy: 'user1',
-        createdTimestamp: '2022-02-14T10:00:00',
-        modifiedTimestamp: '2022-02-14T10:05:00',
-      }
-
-      visitSchedulerApiClient.changeBookedVisit.mockResolvedValue(returnedVisit)
-      whereaboutsApiClient.getEvents.mockResolvedValue([])
-
-      const result = await visitSessionsService.changeBookedVisit({
-        username: 'user',
-        visitSessionData,
-      })
-
-      expect(visitSchedulerApiClient.changeBookedVisit).toHaveBeenCalledTimes(1)
-      expect(result).toEqual(returnedVisit)
-    })
-  })
-
-  describe('cancelVisit', () => {
-    it('should cancel a visit, giving the status code and reason', async () => {
-      const expectedResult: Visit = {
-        applicationReference: 'aaa-bbb-ccc',
-        reference: 'ab-cd-ef-gh',
-        prisonerId: 'AF34567G',
-        prisonId: 'HEI',
-        sessionTemplateReference: 'v9d.7ed.7u',
-        visitRoom: 'A1 L3',
-        visitType: 'SOCIAL',
-        visitStatus: 'CANCELLED',
-        visitRestriction: 'OPEN',
-        startTimestamp: '2022-02-14T10:00:00',
-        endTimestamp: '2022-02-14T11:00:00',
-        visitNotes: [
-          {
-            type: 'VISIT_OUTCOMES',
-            text: 'VISITOR_CANCELLED',
-          },
-          {
-            type: 'STATUS_CHANGED_REASON',
-            text: 'cancellation reason',
-          },
-        ],
-        visitors: [
-          {
-            nomisPersonId: 1234,
-          },
-        ],
-        visitorSupport: [],
-        createdBy: 'user1',
-        createdTimestamp: '2022-02-14T10:00:00',
-        modifiedTimestamp: '2022-02-14T10:05:00',
-      }
-
-      const outcome: OutcomeDto = {
-        outcomeStatus: 'VISITOR_CANCELLED',
-        text: 'cancellation reason',
-      }
-
-      visitSchedulerApiClient.cancelVisit.mockResolvedValue(expectedResult)
-      const result = await visitSessionsService.cancelVisit({
-        username: 'user',
-        reference: expectedResult.reference,
-        outcome,
-      })
-
-      expect(visitSchedulerApiClient.cancelVisit).toHaveBeenCalledTimes(1)
-      expect(visitSchedulerApiClient.cancelVisit).toHaveBeenCalledWith(expectedResult.reference, outcome)
-      expect(result).toEqual(expectedResult)
     })
   })
